Added a new chapter about CD ripping (only in German version)
chapter4.tex now contains a tutorial how to rip a CD using the command line
software abcde. This is working on a Linux PC and also on the Raspiblaster.

This topic was added only in the German version (raspiblaster_de.tex) as a
first step. A translation into English language will follow soon...